+++ src/main/java/dev/chha/incidenttracker/repositories/IncidentRepository.java
@@ -3,5 +3,5 @@ package dev.chha.incidenttracker.repositories;
 import dev.chha.incidenttracker.entities.Incident;
 import org.springframework.data.repository.CrudRepository;
 
-public interface IncidentRepository extends CrudRepository<Incident, Integer> {
+public interface IncidentRepository extends CrudRepository<Incident, Long> {
 }
